reorganize public health analytics files
