+++ app/javascript/components/analytics/display/CustomizedAxisTick.js
