+++ app/javascript/components/analytics/public_health/maps/mapData.js
